+++ frontend/src/types/generated.ts
@@ -1072,10 +1072,11 @@ export type CredentialsUpdated = {
 export type DataRegistered = {
   __typename?: 'DataRegistered';
   dataAsHex: Scalars['String'];
+  decoded: DecodedText;
 };
 
-export type DecodedTransferMemo = {
-  __typename?: 'DecodedTransferMemo';
+export type DecodedText = {
+  __typename?: 'DecodedText';
   decodeType: TextDecodeType;
   text: Scalars['String'];
 };
@@ -2478,7 +2479,7 @@ export type TransactionsEdge = {
 
 export type TransferMemo = {
   __typename?: 'TransferMemo';
-  decoded: DecodedTransferMemo;
+  decoded: DecodedText;
   rawHex: Scalars['String'];
 };