+++ backend/CHANGELOG.md
@@ -7,5 +7,6 @@
 - Added query to fetch backend version.
 - Sorting on Nodes Page
 - Added Accounts Updated GraphQl subscription. Can be used to subscribe to updates of a specific Account. 
+- Fixed Issue 5 by considering 0 balance for Newly created accounts.
 - Added boolean filters on Bakers Query to enable filtering on Baker Removed State.
 - Fixed Memo CBOR Decoding.